move Firstrun Fragment under rocket.firstrun
